+++ Gemfile
@@ -43,6 +43,8 @@ gem "dotenv-rails"
 gem "aws-sdk-s3", require: false
 gem 'kaminari'
 gem 'sendgrid-ruby'
+gem 'jquery-rails'
+gem 'jquery-turbolinks'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
