+++ app/assets/stylesheets/application.scss
@@ -19,10 +19,60 @@
 
 @charset "utf-8";
 
+.slick-list {
+  margin: 0 auto;
+  text-align: center;
+}
+
+.logo_img {
+  width: 130px;
+  margin: 1rem;
+}
+
 .logo {
   margin: 0 auto;
   text-align: center;
 }
+
+.container {
+  margin: 0 auto;
+}
+
+.search_container {
+  box-sizing: border-box;
+  position: relative;
+  border: 1px solid #999;
+  padding: 3px 10px;
+  border-radius: 20px;
+  height: 2.3em;
+  width: 260px;
+  overflow: hidden;
+}
+
+.search_container input[type="text"] {
+  border: none;
+  height: 2.0em;
+}
+
+.search_container input[type="text"]:focus {
+  outline: 0;
+}
+
+.search_container input[type="submit"] {
+  cursor: pointer;
+  font-family: FontAwesome;
+  font-size: 1.3em;
+  border: none;
+  background: none;
+  color: #3879D9;
+  position: absolute;
+  width: 2.5em;
+  height: 2.5em;
+  right: 0;
+  top: -10px;
+  outline: none;
+}
+
 
 /*========= 途中からハンバーガーメニューに変化するのためのCSS ===============*/
 /*========= ボタンのためのCSS ===============*/
@@ -201,4 +251,45 @@
   letter-spacing: 0.1em;
   font-weight: bold;
   transition: all 0.3s;
+}
+
+
+footer {
+  color: #808080;
+  background: #e5e5e5;
+  text-align: center;
+  padding: 30px;
+}
+
+footer a {
+  color: #808080;
+  text-decoration: none;
+}
+
+footer a:hover {
+  text-decoration: underline;
+}
+
+footer .menu {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+footer .menu li {
+  margin: 0;
+  padding: 0 20px;
+  border-right: 1px #808080 solid;
+}
+
+footer .menu li:last-child {
+  border: none;
+}
+
+footer .copyright {
+  margin: 0;
+  padding: 20px 0 0 0;
 }
