+++ app/views/foods/_food.html.erb
@@ -10,7 +10,7 @@
 <% if current_user&.admin %>
   <td><%= link_to 'Edit', edit_food_path(food) %></td>
   <td><%= link_to 'Destroy', food_path(food), method: :delete, data: { confirm: 'Are you sure?' } %></td>
-  <% end %>
+<% end %>
 
 <% if !current_user%>
 <td>ログインしていません。</td>
@@ -24,7 +24,7 @@
 <% else %>
   <td>
     <%= link_to create_favorite_path(food), class: "favorite-link", method: :POST, remote: true do %>
-      いいねをする　 <i class="fa fa-heart like-btn fa-lg"></i> 
+      いいねをする　 <i class="fa fa-heart like-btn fa-lg"></i>
     <% end %>
     <%= food.favorites.count %>
   </td>
