+++ app/views/foods/index.html.erb
@@ -1,19 +1,19 @@
 <p id="notice"><%= notice %></p>
+<div class="slick-list"><img class="top_main" src="/assets/main.png" alt=""></div>
 
 <%= search_form_for @q do |f| %>
-  <%= f.label :item_name_or_variety_or_comment_or_storage_method_or_harvest_or_price_cont, "Keyword" %>
+  <%= f.label :item_name_or_variety_or_comment_or_storage_method_or_harvest_or_price_cont, "キーワード検索" %>
   <%= f.search_field :item_name_or_variety_or_comment_or_storage_method_or_harvest_or_price_cont %>
   <%= f.submit "検索" %>
 <% end %>
-<h1>Foods</h1>
 
 <table>
-  <thead>
+  <%# <thead>
     <tr>
       <th>item_name</th>
       <th colspan="2"></th>
     </tr>
-  </thead>
+  </thead> %>
   <tbody>
     <% @foods.each do |food| %>
       <tr id="food_<%= food.id %>">
