+++ app/views/layouts/application.html.erb
@@ -13,34 +13,49 @@
 
   <body>
     <header id="header">
-      <h1>Logo</h1>
+      <h1></h1>
         <nav id="g-navi">
           <ul>
-            <li><a href="#area-1">ログイン</a></li>
-            <li><a href="#area-2">アクセス</a></li>
-            <li><a href="#area-3">お問い合わせ</a></li>
+            <% if user_signed_in? %>
+            <li><%= link_to "マイページ", users_show_path %></li>
+            <li><%= link_to "ログアウト", destroy_user_session_path, method: :delete %></li>
+            <li><%= link_to '内山農園について', others_path %></li>
+            <li><%= link_to 'お問い合わせ', new_contact_path %></li>
+              <% else %>
+              <li><%= link_to "新規登録", new_user_registration_path %></li>
+              <li><%= link_to "ログイン", new_user_session_path %></li>
+              <li><%= link_to 'ゲストログイン（閲覧用）', users_guest_sign_in_path, method: :post %></li>
+              <li><%= link_to '管理者ログイン（閲覧用）', users_admin_sign_in_path, method: :post %></li>
+              <li><%= link_to '内山農園について', others_path %></li>
+              <li><%= link_to 'お問い合わせ', new_contact_path %></li>
+            <% end %>
           </ul>
         </nav>
     </header>
-    <div class="logo"><img src="/assets/uchiyamafarmlogo.png" alt=""></div>
+    <div class="logo"><%= link_to (foods_path) do%><img class="logo_img" src="/assets/uchiyamafarmlogo.png" alt=""><% end%></div>
 
-    <% if user_signed_in? %>
-      <%= link_to "ログアウト", destroy_user_session_path, method: :delete %>
-    <% else %>
-      <%= link_to "新規登録", new_user_registration_path %>
-      <%= link_to "ログイン", new_user_session_path %>
-      <%= link_to 'ゲストログイン（閲覧用）', users_guest_sign_in_path, method: :post %>
-      <%= link_to '管理者ログイン（閲覧用）', users_admin_sign_in_path, method: :post %>
-    <% end %>
-    <p><i class="fa-light env-btn fa-envelope"></i></p>
+    <div class="container">
+      <% if notice %>
+        <p class="alert alert-notice"><%= notice %></p>
+      <% end %>
+      <% if alert %>
+        <p class="alert alert-error"><%= alert %></p>
+      <% end %>
 
-    <% if notice %>
-      <p class="alert alert-notice"><%= notice %></p>
-    <% end %>
-    <% if alert %>
-      <p class="alert alert-error"><%= alert %></p>
-    <% end %>
-    <%= yield %>
+      <%= yield %>
+    </div>
+    <footer>
+      <ul class="menu">
+        <li><a href="#">AAAAA</a></li>
+        <li><a href="#">BBBBB</a></li>
+        <li><a href="#">CCCCC</a></li>
+        <li><a href="#">DDDDD</a></li>
+        <li><a href="#">EEEEE</a></li>
+      </ul>
+      <p class="copyright">
+        &copy; My Site
+      </p>
+    </footer>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
     <script src="https://code.jquery.com/jquery-3.4.1.min.js" integrity="sha256-CSXorXvZcTkaix6Yvo6HppcZGetbYMGWSFlBw8HfCJo=" crossorigin="anonymous"></script>
   </body>
