+++ app/views/plans/_form.html.erb
@@ -1,0 +1,1 @@
+<p id="notice"><%= notice %></p>
