+++ app/views/plans/new.erb
@@ -1,0 +1,14 @@
+<% if @plan.errors.any? %>
+  <div id="error_explanation">
+    <h2><%= pluralize(@plan.errors.count, "error") %> prohibited this <%= @plan.title %> from being saved:</h2>
+    <ul>
+    <% @plan.errors.full_messages.each do |message| %>
+      <li><%= message %></li>
+    <% end %>
+    </ul>
+  </div>
+<% end %>
+
+<%= render 'form', plan: @plan %>
+
+<%= link_to 'Back', plans_path %>
