+++ package.json
@@ -7,6 +7,7 @@
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "5.4.3",
+    "jquery": "^3.6.0",
     "rails_admin": "3.0.0",
     "turbolinks": "^5.2.0",
     "webpack": "^4.46.0",
